Ajout des solutions
